+++ src/collective/formsupport/counter/adapters/post.py
@@ -9,7 +9,6 @@ from copy import deepcopy
 from plone import api
 from zope.annotation.interfaces import IAnnotations
 from zope.component import adapter
-from zope.i18n import translate
 from zope.interface import implementer
 from zope.interface import Interface
 
@@ -17,7 +16,6 @@ from zope.interface import Interface
 @implementer(IPostAdapter)
 @adapter(Interface, ICollectiveFormsupportCounterLayer)
 class PostAdapterWithCounter(PostAdapter):
-
     _block = {}
 
     @property
@@ -47,11 +45,12 @@ class PostAdapterWithCounter(PostAdapter):
         annotations = IAnnotations(self.context)
 
         value = annotations.get(COUNTER_ANNOTATIONS_NAME, {}).get(block_id, 0) + 1
+
         form_data["data"].append(
             {
                 "field_id": COUNTER_BLOCKS_FIELD_ID,
-                "label": translate(
-                    _("Form counter"), target_language=api.portal.get_current_language()
+                "label": api.portal.translate(
+                    _("form_counter_label", default="Counter")
                 ),
                 "value": value,
             }