+++ src/collective/formsupport/counter/datamanager/__init__.py
@@ -1,10 +1,11 @@
+from collective.formsupport.counter import _
 from collective.formsupport.counter.config import COUNTER_BLOCKS_FIELD_ID
 from collective.formsupport.counter.config import COUNTER_ENABLED_FORM_FLAG_NAME
 from collective.formsupport.counter.interfaces import ICollectiveFormsupportCounterLayer
 from collective.volto.formsupport.datamanager.catalog import FormDataStore
 from collective.volto.formsupport.interfaces import IFormDataStore
 from collective.volto.formsupport.utils import get_blocks
-from copy import deepcopy
+from plone import api
 from plone.dexterity.interfaces import IDexterityContent
 from zope.component import adapter
 from zope.interface import implementer
@@ -13,35 +14,35 @@ from zope.interface import implementer
 @implementer(IFormDataStore)
 @adapter(IDexterityContent, ICollectiveFormsupportCounterLayer)
 class FormDataStoreWithCounter(FormDataStore):
-    def get_form_fields(self):
+    def get_form_block(self):
         blocks = get_blocks(self.context)
-
         if not blocks:
             return {}
-
-        form_block = {}
-
         for id, block in blocks.items():
-
             if id != self.block_id:
                 continue
-
             block_type = block.get("@type", "")
-
             if block_type == "form":
-                form_block = deepcopy(block)
+                return block
+        return {}
 
-        if not form_block:
-            return {}
+    def get_form_fields(self):
+        """
+        Patch: add form_counter to fields list
+        """
+        subblocks = super().get_form_fields()
+        if not subblocks:
+            return subblocks
 
-        subblocks = form_block.get("subblocks", [])
-
+        form_block = self.get_form_block()
         if form_block.get(COUNTER_ENABLED_FORM_FLAG_NAME):
-            subblocks.append({"field_id": COUNTER_BLOCKS_FIELD_ID})
-
-        # Add the 'custom_field_id' field back in as this isn't stored with each subblock
-        for index, field in enumerate(subblocks):
-            if form_block.get(field["field_id"]):
-                subblocks[index]["custom_field_id"] = form_block.get(field["field_id"])
+            subblocks.append(
+                {
+                    "field_id": COUNTER_BLOCKS_FIELD_ID,
+                    "label": api.portal.translate(
+                        _("form_counter_label", default="Counter")
+                    ),
+                }
+            )
 
         return subblocks